Change handling metadata and website content
Markdown files for articles are now replaced with directories containing a markdown file and a metadata.json file.

Support for YAML syntax is limited in Flexmark and Kaml (a yaml serialization library for Kotlin). Standard JSON syntax allows arbitrary nesting of objects in JSON file and makes it possible to use standard deserialization available in kotlinx.serialization.

Additionally, separating page metadata from content makes it easier to parse both when a front-matter format is not already available as an extension for flexmark and makes it easier to implement content formats other than markdown in the future, because not all formats support including metadata as a frontmatter section within the same file.